Add dummy file_open method
This method will be subclassed/defined by extensions and be called every
time a file is opened by Nautilus.

A similar method, update_file_info, is called by Nautilus each time a
file is displayed on screen. This method can be used by extensions in
order to change/update information about those files such as emblems or
descriptions.

file_open will be called whenever a file is set to open, and depending
on whether the extension returns a True/False for the method, Nautilus
will or will not open the file.